Move preference management of reader related preferences to a separate class.
